(f417b026f) Fetched: Changes for playing video tutorial from local branch

+++ Barotrauma/BarotraumaServer/Source/Networking/Client.cs
@@ -46,7 +46,7 @@ namespace Barotrauma.Networking
         
         //when was a specific entity event last sent to the client
         //  key = event id, value = NetTime.Now when sending
-        public readonly Dictionary<UInt16, double> EntityEventLastSent = new Dictionary<UInt16, double>();
+        public readonly Dictionary<UInt16, float> EntityEventLastSent = new Dictionary<UInt16, float>();
         
         //when was a position update for a given entity last sent to the client
         //  key = entity id, value = NetTime.Now when sending
